+++ core/src/ldml/ldml_transforms.hpp
@@ -33,18 +33,6 @@ namespace km {
 namespace core {
 namespace ldml {
 
-/** @returns true on success */
-inline bool uassert_success(const char *file, int line, const char *function, UErrorCode status) {
-  if (U_FAILURE(status)) {
-    DebugLog2(file, line, function, "U_FAILURE(%s)", u_errorName(status));
-    return false;
-  } else {
-    return true;
-  }
-}
-
-/** the first assert is for debug builds, the second triggers the debuglog and has the return value. */
-#define UASSERT_SUCCESS(status) (assert(U_SUCCESS(status)), uassert_success(__FILE__, __LINE__, __FUNCTION__, status))
 
 using km::core::kmx::SimpleUSet;
 
@@ -297,85 +285,6 @@ public:
        const core::kmx::COMP_KMXPLUS_TRAN *tran,
        const core::kmx::COMP_KMXPLUS_TRAN_Helper &tranHelper);
 };
-
-// string routines
-
-/** indicates that the marker was before the end of text. */
-const char32_t MARKER_BEFORE_EOT = km::core::kmx::Uni_FFFE_NONCHARACTER;
-
-/** specify the type of encoding for marker text */
-enum marker_encoding {
-  /** encoding as UC_SENTINEL + CODE_DEADKEY + <number> */
-  plain_sentinel,
-  /** encoding as a regex matching the marker */
-  regex_sentinel,
-};
-
-/** a marker ID (1-based) */
-typedef KMX_DWORD marker_num;
-
-/** list of markers */
-typedef std::deque<marker_num> marker_list;
-
-/** map from following-char to marker numbers. */
-typedef std::map<char32_t, marker_list> marker_map;
-
-/** Normalize a u32string inplace to NFD. @return false on failure */
-bool normalize_nfd(std::u32string &str);
-/** Normalize a u16string inplace to NFD. @return false on failure */
-bool normalize_nfd(std::u16string &str);
-/** Normalize a u32string inplace to NFD, retaining markers.
- * @param markers will be populated with marker chars
- * @return false on failure
- **/
-bool normalize_nfd_markers_segment(std::u32string &str, marker_map &markers, marker_encoding encoding = plain_sentinel);
-bool normalize_nfd_markers_segment(std::u16string &str, marker_map &markers, marker_encoding encoding = plain_sentinel);
-bool normalize_nfd_markers(std::u32string &str, marker_encoding encoding = plain_sentinel);
-bool normalize_nfd_markers(std::u16string &str, marker_encoding encoding = plain_sentinel);
-
-// /** Normalize a u32string inplace to NFC, retaining markers.
-//  * @param markers will be populated with marker chars
-//  * @return false on failure
-//  **/
-// bool normalize_nfd_markers_segment(std::u32string &str, marker_map &markers, marker_encoding encoding = plain_sentinel);
-// bool normalize_nfd_markers_segment(std::u16string &str, marker_map &markers, marker_encoding encoding = plain_sentinel);
-// inline bool normalize_nfc_markers(std::u32string &str, marker_encoding encoding = plain_sentinel);
-// inline bool normalize_nfc_markers(std::u16string &str, marker_encoding encoding = plain_sentinel);
-
-// /** Normalize a u32string inplace to NFC. @return false on failure */
-// bool normalize_nfc(std::u32string &str);
-
-// /** Normalize a u16string inplace to NFC. @return false on failure */
-// bool normalize_nfc(std::u16string &str);
-
-/** Remove markers and optionally note their glue characters in the map */
-std::u32string remove_markers(const std::u32string &str, marker_map *markers = nullptr, marker_encoding encoding = plain_sentinel);
-
-/** same but with a reference */
-inline std::u32string remove_markers(const std::u32string &str, marker_map &markers, marker_encoding encoding = plain_sentinel)  {
-  return remove_markers(str, &markers, encoding);
-}
-
-/** prepend the marker string in UC_SENTINEL format to the str */
-void prepend_marker(std::u32string &str, marker_num marker, marker_encoding encoding = plain_sentinel);
-
-/** format 'marker' as 0001...FFFF and put it at the beginning of the string */
-void prepend_hex_quad(std::u32string &str, KMX_DWORD marker);
-
-/** parse 0001...FFFF into a KMX_DWORD. Returns 0 on failure */
-KMX_DWORD parse_hex_quad(const km_core_usv hex_str[]);
-
-
-// bool normalize_nfc_markers(std::u16string &str, marker_encoding encoding) {
-//   marker_map m;
-//   return normalize_nfc_markers_segment(str, m, encoding);
-// }
-
-
-// bool normalize_nfc_markers(std::u32string &str, marker_encoding encoding) {
-//   marker_map m;
-//   return normalize_nfc_markers_segment(str, m, encoding);
-// }
 
 
 }  // namespace ldml
