+++ core/tests/unit/ldml/ldml.cpp
@@ -29,7 +29,7 @@
 #include "ldml_test_source.hpp"
 #include "debuglog.h"
 
-#include "ldml/ldml_transforms.hpp"
+#include "ldml/ldml_markers.hpp"
 
 namespace {
