+++ core/tests/unit/ldml/ldml_test_source.cpp
@@ -24,7 +24,7 @@
 #include <kmx/kmx_plus.h>
 #include "ldml/keyman_core_ldml.h"
 #include "ldml/ldml_processor.hpp"
-#include "ldml/ldml_transforms.hpp"
+#include "ldml/ldml_markers.hpp"
 
 #include "path.hpp"
 #include "state.hpp"
