+++ core/tests/unit/ldml/test_transforms.cpp
@@ -1,4 +1,5 @@
 #include "../../../src/ldml/ldml_transforms.hpp"
+#include "../../../src/ldml/ldml_markers.hpp"
 #include "kmx/kmx_plus.h"
 #include "kmx/kmx_xstring.h"
 #include <iostream>
